updated signposterboard:
renamed to signboard.cs
cleaned up the script
adjusted other scripts using signboard.cs (such as arenabuilders.cs)
